add documents

+++ shield/views/document.py
@@ -1,4 +1,6 @@
-from django.http import HttpResponseRedirect
+import os
+from django.conf import settings
+from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render, redirect
 from django.views import View
 from django.views.generic import (
@@ -18,6 +20,9 @@ from django.core.exceptions import ValidationError
 
 from shield.forms.document import DocumentForm
 from shield.models.document import Document
+from shield.models.folder import Folder
+from django.http import FileResponse, Http404
+from django.shortcuts import get_object_or_404
 
 
 class DocumentListView(ListView):
@@ -30,14 +35,30 @@ class DocumentListView(ListView):
         return context
 
 
-class DocumentCreateView(SuccessMessageMixin, CreateView):
-    model = Document
-    form_class = DocumentForm
-    template_name = "document/create.html"
-    success_message = "document created successfully"
+class DocumentCreateView(SuccessMessageMixin, View):
+    def get(self, request, pk):
+        folder = get_object_or_404(Folder, pk=pk)
+        form = DocumentForm()
+        return render(request, "document/create.html", {"folder": folder, "form": form })
+    
+    def post(self, request, pk):
+        folder = get_object_or_404(Folder, pk=pk)
+        form = DocumentForm(request.POST, request.FILES)
+        if form.is_valid():          
+            document = form.save(commit=False)
+            document.save()
 
-    def get_success_url(self):
-        return reverse("document-index")
+            if document:
+                messages.success(request,'Document Successfully')
+                return redirect('folder-details', pk = folder.pk )
+            else:
+                messages.warning(request,'Error placing bid')
+                return redirect(request.META.get('HTTP_REFERER', '/'))
+        else:
+            messages.warning(request, form.errors)
+            return render(request, "document/create.html", {"folder": folder, "form": form})
+
+
 
 class DocumentDetailsView(DetailView):
     model = Document
@@ -59,9 +80,28 @@ class DocumentUpdateView(SuccessMessageMixin, UpdateView):
 class DocumentDeleteView(View):
     def get(self, request, **kwargs):
         obj = get_object_or_404(Document, pk=kwargs.get('pk'))
-        obj.delete()
-        messages.success(request,f'{obj} deleted successfully')
-        return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
-            
+        if request.user not in obj.folder.users.all():
+            messages.warning(request,f'You are not authorized to delete this document')
+            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+        
+        else:
+            obj.delete()
+            messages.success(request,f'{obj} deleted successfully')
+            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
 
+class DocumentDownloadView(View):
+    def get(self, request, pk):
+        # Retrieve the document object
+        document = get_object_or_404(Document, pk=pk)
 
+        # Check if the user has permission to download the document
+        if request.user not in document.folder.users.all():
+            messages.warning(request,f'You are not authorized to download this document')
+            return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+
+        try:
+            file_path = document.path.path
+            response = FileResponse(open(file_path, 'rb'), as_attachment=True)
+            return response
+        except FileNotFoundError:
+            raise Http404
